Also pull model spec up to standardize across registry implementations
